+++ Input/Input.js
@@ -303,6 +303,11 @@ const InputBase = kind({
 
 	render: ({clearButton, clearIcon, css, dir, disabled, handleClear, iconAfter, iconBefore, invalidTooltip, onChange, placeholder, size, type, value, ...rest}) => {
 		const inputProps = extractInputProps(rest);
+
+		if (type === 'password') {
+			inputProps.spellCheck = false;
+		}
+
 		delete rest.dismissOnEnter;
 		delete rest.focused;
 		delete rest.invalid;