+++ samples/qa-scroller/src/components/Controls/Controls.js
@@ -9,29 +9,21 @@ import ScrollModeSwitch from '../ScrollModeSwitch';
 
 const Controls = ({handleFocusableScrollbar, handleHeight, handleScrollMode, handleWidth, height, nativeScroll, width}) => {
 	const inputWidth = {width: '5em'};
+	const rowWidth = typeof window !== 'undefined' ? `${ri.scaleToRem(window.innerWidth)}` : `${ri.scaleToRem(1920)}`;
 
 	return (
-		<Row style={{width: `${ri.scale(window.innerWidth)}px`}}>
-			<Cell>
+		<Row style={{width: rowWidth}} wrap>
+			<Cell align="center" shrink>
 				<label>height:</label>
 				<Input onChange={handleHeight} size="small" style={inputWidth} type="number" value={height} />
 			</Cell>
-			<Cell>
+			<Cell align="center" shrink>
 				<label>width:</label>
 				<Input onChange={handleWidth} size="small" style={inputWidth} type="number" value={width} />
 			</Cell>
-			<Cell
-				component={CheckboxItem}
-				onClick={handleFocusableScrollbar}
-			>
-				Focusable Scrollbar
-			</Cell>
-			<Cell>
-				<ScrollModeSwitch defaultSelected={nativeScroll} onToggle={handleScrollMode} />
-			</Cell>
-			<Cell>
-				<LocaleSwitch />
-			</Cell>
+			<CheckboxItem onClick={handleFocusableScrollbar} style={{minWidth: '12em'}}>Focusable Scrollbar</CheckboxItem>
+			<ScrollModeSwitch defaultSelected={nativeScroll} onToggle={handleScrollMode} style={{minWidth: '12em'}} />
+			<LocaleSwitch style={{minWidth: '12em'}} />
 		</Row>
 	);
 };